requirements.txt
aiodns==3.2.0
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
# aiohttp[speedups]
aiohttp[speedups]==3.9.5 # for windows use this instead
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
babel==2.16.0
black==24.10.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
flake8==7.1.1
fluent.runtime==0.4.0
fluent.syntax==0.19.0
frozenlist==1.5.0
idna==3.10
inflect==7.4.0
isort==5.13.2
magic-filter==1.0.12
mccabe==0.7.0
more-itertools==10.5.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.0
pycares==4.4.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
pytz==2024.2
ruff==0.8.1
structlog==24.4.0
typeguard==4.4.0
typing_extensions==4.12.2
ujson==5.10.0
yarl==1.17.0
